Improvements based on ShellCheck
- Use shebang even for sourced snippets of shell code
- Remove dead parameters
- Cleaner command return value check; also allows us to remove ugly boilerplate code for errexit disable/enable
- All file paths with leading globs now use the ./ prefix to avoid potential interpretation of files as flags
- Array assignment now use recommended mapfile builtin
- Whenever a subshell expression is evaluated, declare variable ahead of assignment
- All calls to find now use . explicitly to avoid potential issues
- Only execute rm -rf on a path when the variable exists (note: should not have been possible before anyhow due to preserve-root behavior on modern systems)
- Strictly avoid read-write in the same pipeline, even for command substitution (which should be safe, but not aware of documentation on this)
- Rename temporary link file to factory-image-link to avoid confusion for ShellCheck
- Add missing quoting where expansion and splitting is not desired

+++ scripts/shared/analysis/20_diffoscope-files.sh
@@ -20,33 +20,27 @@ function preProcessImage {
     # Mutable diffoscope params
     local DIFF_IN_META="$1"
 
-    local DIFF_IN_BASE=$(eval echo \$"$DIFF_IN_META")
+    local DIFF_IN_BASE
+    DIFF_IN_BASE=$(eval echo \$"$DIFF_IN_META")
     local DIFF_IN_RESOLVED=$DIFF_IN_BASE
     # Sanity check that we are dealing with a image
-    set +o errexit # Disable early exit
-    file "${DIFF_IN_RESOLVED}" | grep -P '(ext2)|(ext3)|(ext4)|(Android sparse image)'
-    if [[ "$?" -eq 0 ]]; then
-        set -o errexit # Re-enable early exit
-
+    if file "${DIFF_IN_RESOLVED}" | grep -P '(ext2)|(ext3)|(ext4)|(Android sparse image)'; then
         # Detect sparse images
-        set +o errexit # Disable early exit
-        file "${DIFF_IN_RESOLVED}" | grep 'Android sparse image'
-        if [[ "$?" -eq 0 ]]; then
-            set -o errexit # Re-enable early exit
+        if file "${DIFF_IN_RESOLVED}" | grep 'Android sparse image'; then
             # Deomcpress into raw ext2/3/4 partition image
             "${AOSP_HOST_BIN}/simg2img" "${DIFF_IN_RESOLVED}" "${DIFF_IN_RESOLVED}.raw"
-            eval $DIFF_IN_META="${DIFF_IN_RESOLVED}.raw"
+            eval "$DIFF_IN_META=${DIFF_IN_RESOLVED}.raw"
         fi
-        set -o errexit # Re-enable early exit
 
-        local DIFF_IN_RESOLVED=$(eval echo \$"$DIFF_IN_META")
+        local DIFF_IN_RESOLVED
+        DIFF_IN_RESOLVED=$(eval echo \$"$DIFF_IN_META")
         # Mount image to ensure stable file iteration order
         mkdir -p "${DIFF_IN_RESOLVED}.mount"
         # guestfs allows working with complex file systems images (e.g. LVM volume groups with multiple LVs) and thus requires
         # us to be explicit about what part of image we want to mount. Running `virt-filesystems -a system.img.raw` returns the
         # name of the "virtual" device and can be directly fed to guestmount (usually this is /dev/sda)
         guestmount -o "uid=$(id -u)" -o "gid=$(id -g)" -a "${DIFF_IN_RESOLVED}" -m "$(virt-filesystems -a "${DIFF_IN_RESOLVED}")" --ro "${DIFF_IN_RESOLVED}.mount"
-        eval $DIFF_IN_META="${DIFF_IN_RESOLVED}.mount"
+        eval "$DIFF_IN_META=${DIFF_IN_RESOLVED}.mount"
 
         # Extract apex_payload.img from APEX archives for separate diffoscope run
         if [[ "$(find "${DIFF_IN_RESOLVED}.mount" -type f -iname '*.apex' | wc -l)" -ne 0 ]]; then
@@ -63,35 +57,31 @@ function preProcessImage {
             # Perform linking for these to the common prefix to enable filename based matching
             (
                 cd "${DIFF_IN_BASE}.apexes" && \
-                find -type f -iname 'com.google.android.*-apex_payload.img' \
+                find . -type f -iname 'com.google.android.*-apex_payload.img' \
                     -exec bash -c 'ln -s "$0" "$(echo "$0" | sed "s/com.google.android/com.android/")"' "{}" \;
             )
             # AFAIK the bash -c invokation is needed to make the sed subshell invocation lazily evaluated during find result iteration
 
-            # Have another look at the list if files in common, but only consider APEX related ones
+            # Have another look at the list of files in common, but only consider APEX related ones
             local -r APEX_FOLDER_BASENAME="$(basename "${DIFF_IN_BASE}.apexes")"
-            local -ar APEX_PAYLOAD_FILES=($(comm -12 \
+            local -a APEX_PAYLOAD_FILES
+            mapfile -t APEX_PAYLOAD_FILES < <(comm -12 \
                 <(cd "${IN_DIR_1}" && find "${APEX_FOLDER_BASENAME}" -type 'f,l' | sort) \
                 <(cd "${IN_DIR_2}" && find "${APEX_FOLDER_BASENAME}" -type 'f,l' | sort) \
-            ))
+            )
+            declare -r APEX_PAYLOAD_FILES
             # Append to list of files requiring processing via diffoscope
             FILES+=( "${APEX_PAYLOAD_FILES[@]}" )
         fi
     fi
-    set -o errexit # Re-enable early exit
-
 }
 
 function postProcessImage {
     local DIFF_IN="$1"
 
     # Sanity check that we are dealing with a mount
-    set +o errexit # Disable early exit
-    mount | grep "${DIFF_IN}"
-    if [[ "$?" -eq 0 ]]; then
-        set -o errexit # Re-enable early exit
-
-        IMAGE="$(dirname $DIFF_IN)/$(basename -s '.mount' "$DIFF_IN")"
+    if mount | grep "${DIFF_IN}"; then
+        IMAGE="$(dirname "$DIFF_IN")/$(basename -s '.mount' "$DIFF_IN")"
 
         guestunmount "${IMAGE}.mount"
         rmdir "${IMAGE}.mount"
@@ -101,7 +91,6 @@ function postProcessImage {
             rm "$IMAGE"
         fi
     fi
-    set -o errexit # Re-enable early exit
 }
 
 function diffoscopeFile {
@@ -169,7 +158,7 @@ main() {
     # Misc variables + ensure ${OUT_DIR} exists
     local -r AOSP_HOST_BIN="${RB_AOSP_BASE}/src/out/host/linux-x86/bin"
     mkdir -p "${OUT_DIR}"
-    rm -rf "${OUT_DIR}/"* # Clean up previous diff results
+    rm -rf "${OUT_DIR:?}/"* # Clean up previous diff results
 
     # apktool quirk workaround, see https://github.com/iBotPeaches/Apktool/issues/2048
     sudo mkdir -p  "/root/.local/share/apktool/framework"
@@ -179,10 +168,11 @@ main() {
     sudo chmod +r "/boot/vmlinuz-"*
 
     # Create list of files in common for both directories. Ignore super.img, we unpacked it previously
-    local -a FILES=($(comm -12 \
-        <(cd "${IN_DIR_1}" && find -type f | sort) \
-        <(cd "${IN_DIR_2}" && find -type f | sort) \
-    | grep -v 'super.img'))
+    local -a FILES
+    mapfile -t FILES < <(comm -12 \
+        <(cd "${IN_DIR_1}" && find . -type f | sort) \
+        <(cd "${IN_DIR_2}" && find . -type f | sort) \
+    | grep -v 'super.img')
 
     for ((i = 0; i < "${#FILES[@]}"; i++)); do
         FILE="${FILES[$i]}"
